requirements.txt
ultralytics
torch
opencv-python
fastapi
uvicorn
numpy
pydantic
pyyaml
python-multipart
jinja2
